Load plotly as a module on metrics page
