+++ src/Aspire.Dashboard/Components/App.razor
@@ -28,7 +28,5 @@
     <Routes @rendermode="@(new InteractiveServerRenderMode(prerender: false))" />
     <script src="_framework/blazor.web.js"></script>
     <script src="js/app.js"></script>
-    <script src="js/theme.js" type="module"></script>
-    <script src="js/plotly-2.32.0.min.js"></script>
 </body>
 </html>
