+++ src/Aspire.Dashboard/Components/Controls/Chart/PlotlyChart.razor.cs
@@ -16,7 +16,7 @@ using Microsoft.JSInterop;
 
 namespace Aspire.Dashboard.Components;
 
-public partial class PlotlyChart : ChartBase, IDisposable
+public partial class PlotlyChart : ChartBase, IAsyncDisposable
 {
     [Inject]
     public required IJSRuntime JS { get; init; }
@@ -28,6 +28,7 @@ public partial class PlotlyChart : ChartBase, IDisposable
     public required IDialogService DialogService { get; init; }
 
     private DotNetObjectReference<ChartInterop>? _chartInteropReference;
+    private IJSObjectReference? _jsModule;
 
     private string FormatTooltip(string title, double yValue, DateTimeOffset xValue)
     {
@@ -45,6 +46,11 @@ public partial class PlotlyChart : ChartBase, IDisposable
 
     protected override async Task OnChartUpdated(List<ChartTrace> traces, List<DateTimeOffset> xValues, List<ChartExemplar> exemplars, bool tickUpdate, DateTimeOffset inProgressDataTime)
     {
+        if (_jsModule == null)
+        {
+            return;
+        }
+
         var traceDtos = traces.Select(t => new PlotlyTrace
         {
             Name = t.Name,
@@ -71,7 +77,7 @@ public partial class PlotlyChart : ChartBase, IDisposable
             _chartInteropReference?.Dispose();
             _chartInteropReference = DotNetObjectReference.Create(new ChartInterop(this));
 
-            await JS.InvokeVoidAsync("initializeChart",
+            await _jsModule.InvokeVoidAsync("initializeChart",
                 "plotly-chart-container",
                 traceDtos,
                 exemplarTraceDto,
@@ -82,7 +88,7 @@ public partial class PlotlyChart : ChartBase, IDisposable
         }
         else
         {
-            await JS.InvokeVoidAsync("updateChart",
+            await _jsModule.InvokeVoidAsync("updateChart",
                 "plotly-chart-container",
                 traceDtos,
                 exemplarTraceDto,
@@ -154,13 +160,28 @@ public partial class PlotlyChart : ChartBase, IDisposable
         return exemplarTraceDto;
     }
 
-    public void Dispose()
+    protected override async Task OnAfterRenderAsync(bool firstRender)
+    {
+        if (firstRender)
+        {
+            _jsModule = await JS.InvokeAsync<IJSObjectReference>("import", "/js/app-metrics.js");
+        }
+
+        await base.OnAfterRenderAsync(firstRender);
+    }
+
+    // The first data is used to initialize the chart. The module needs to be ready first.
+    protected override bool ReadyForData() => _jsModule != null;
+
+    public async ValueTask DisposeAsync()
     {
         if (_chartInteropReference != null)
         {
             _chartInteropReference.Value.Dispose();
             _chartInteropReference.Dispose();
         }
+
+        await JSInteropHelpers.SafeDisposeAsync(_jsModule);
     }
 
     /// <summary>
