+++ src/Aspire.Dashboard/wwwroot/js/app-metrics.js
@@ -1,0 +1,206 @@
+import './plotly-2.32.0.min.js'
+
+export function initializeChart(id, traces, xValues, rangeStartTime, rangeEndTime, serverLocale) {
+    registerLocale(serverLocale);
+
+    var chartContainerDiv = document.getElementById(id);
+
+    // Reusing a div can create issues with chart lines appearing beyond the end range.
+    // Workaround this issue by replacing the chart div. Ensures we start from a new state.
+    var chartDiv = document.createElement("div");
+    chartContainerDiv.replaceChildren(chartDiv);
+
+    var themeColors = getThemeColors();
+
+    var data = [];
+    for (var i = 0; i < traces.length; i++) {
+        var name = traces[i].name || "Value";
+        var t = {
+            x: traces[i].x,
+            y: traces[i].y,
+            name: name,
+            text: traces[i].tooltips,
+            hoverinfo: 'text',
+            stackgroup: "one"
+        };
+        data.push(t);
+    }
+
+    var points = {
+        x: exemplarTrace.x,
+        y: exemplarTrace.y,
+        name: exemplarTrace.name,
+        text: exemplarTrace.tooltips,
+        hoverinfo: 'text',
+        traceData: exemplarTrace.traceData,
+        mode: 'markers',
+        type: 'scatter',
+        marker: {
+            size: 16,
+            color: themeColors.pointColor,
+            line: {
+                color: themeColors.backgroundColor,
+                width: 1
+            }
+        }
+    };
+    data.push(points);
+
+    // Explicitly set the width and height based on the container div.
+    // If there is no explicit width and height, Plotly will use the rendered container size.
+    // However, if the container isn't visible then it uses a default size.
+    // Being explicit ensures the chart is always the correct size.
+    var width = parseInt(chartContainerDiv.style.width);
+    var height = parseInt(chartContainerDiv.style.height);
+
+    var layout = {
+        width: width,
+        height: height,
+        paper_bgcolor: themeColors.backgroundColor,
+        plot_bgcolor: themeColors.backgroundColor,
+        margin: { t: 0, r: 0, b: 40, l: 50 },
+        xaxis: {
+            type: 'date',
+            range: [rangeEndTime, rangeStartTime],
+            fixedrange: true,
+            tickformat: "%X",
+            color: themeColors.textColor
+        },
+        yaxis: {
+            rangemode: "tozero",
+            fixedrange: true,
+            color: themeColors.textColor
+        },
+        hovermode: "closest",
+        showlegend: true,
+        legend: {
+            orientation: "h",
+            font: {
+                color: themeColors.textColor
+            },
+            traceorder: "normal",
+            itemclick: false,
+            itemdoubleclick: false
+        }
+    };
+
+    var options = { scrollZoom: false, displayModeBar: false };
+
+    Plotly.newPlot(chartDiv, data, layout, options);
+
+    fixTraceLineRendering(chartDiv);
+
+    // We only want a pointer cursor when the mouse is hovering over an exemplar point.
+    // Set the drag layer cursor back to the default and then use plotly_hover/ploty_unhover events to set to pointer.
+    dragLayer = document.getElementsByClassName('nsewdrag')[0];
+    dragLayer.style.cursor = 'default';
+
+    // Use mousedown instead of plotly_click event because plotly_click has issues with updating charts.
+    // The current point is tracked by setting it with hover/unhover events and then mousedown uses the current value.
+    var currentPoint = null;
+    chartDiv.on('plotly_hover', function (data) {
+        var point = data.points[0];
+        if (point.fullData.name == exemplarTrace.name) {
+            currentPoint = point;
+            var pointTraceData = point.data.traceData[point.pointIndex];
+            dragLayer.style.cursor = 'pointer';
+        }
+    });
+    chartDiv.on('plotly_unhover', function (data) {
+        var point = data.points[0];
+        if (point.fullData.name == exemplarTrace.name) {
+            currentPoint = null;
+            dragLayer.style.cursor = 'default';
+        }
+    });
+    chartDiv.addEventListener("mousedown", (e) => {
+        if (currentPoint) {
+            var point = currentPoint;
+            var pointTraceData = point.data.traceData[point.pointIndex];
+
+            chartInterop.invokeMethodAsync('ViewSpan', pointTraceData.traceId, pointTraceData.spanId);
+        }
+    });
+}
+
+export function updateChart(id, traces, xValues, rangeStartTime, rangeEndTime) {
+    var chartContainerDiv = document.getElementById(id);
+    var chartDiv = chartContainerDiv.firstChild;
+
+    var themeColors = getThemeColors();
+
+    var xUpdate = [];
+    var yUpdate = [];
+    var tooltipsUpdate = [];
+    var traceData = [];
+    for (var i = 0; i < traces.length; i++) {
+        xUpdate.push(traces[i].x);
+        yUpdate.push(traces[i].y);
+        tooltipsUpdate.push(traces[i].tooltips);
+        traceData.push(traces.traceData);
+    }
+
+    xUpdate.push(exemplarTrace.x);
+    yUpdate.push(exemplarTrace.y);
+    tooltipsUpdate.push(exemplarTrace.tooltips);
+    traceData.push(exemplarTrace.traceData);
+
+    var data = {
+        x: xUpdate,
+        y: yUpdate,
+        text: tooltipsUpdate,
+        traceData: traceData
+    };
+
+    var layout = {
+        xaxis: {
+            type: 'date',
+            range: [rangeEndTime, rangeStartTime],
+            fixedrange: true,
+            tickformat: "%X",
+            color: themeColors.textColor
+        }
+    };
+
+    Plotly.update(chartDiv, data, layout);
+
+    fixTraceLineRendering(chartDiv);
+}
+
+function getThemeColors() {
+    // Get colors from the current light/dark theme.
+    var style = getComputedStyle(document.body);
+    return {
+        backgroundColor: style.getPropertyValue("--fill-color"),
+        textColor: style.getPropertyValue("--neutral-foreground-rest"),
+        pointColor: style.getPropertyValue("--accent-fill-rest")
+    };
+}
+
+function fixTraceLineRendering(chartDiv) {
+    // In stack area charts Plotly orders traces so the top line area overwrites the line of areas below it.
+    // This isn't the effect we want. When the P50, P90 and P99 values are the same, the line displayed is P99
+    // on the P50 area.
+    //
+    // The fix is to reverse the order of traces so the correct line is on top. There isn't a way to do this
+    // with CSS because SVG doesn't support z-index. Node order is what determines the rendering order.
+    //
+    // https://github.com/plotly/plotly.js/issues/6579
+    var parent = chartDiv.querySelector(".scatterlayer");
+
+    if (parent.childNodes.length > 0) {
+        for (var i = 1; i < parent.childNodes.length; i++) {
+            var child = parent.childNodes[i];
+            parent.insertBefore(child, parent.firstChild);
+        }
+
+        // Check if there is a trace with points. It should be top most.
+        for (var i = 0; i < parent.childNodes.length; i++) {
+            var child = parent.childNodes[i];
+            if (child.querySelector(".point")) {
+                // Append trace to parent to move to the last in the collection.
+                parent.appendChild(child);
+            }
+        }
+    }
+}
